LPD-36028 Move package to osgi.util.tracker
