+++ modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/osgi/util/tracker/IndexOnStartupExecutor.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.search.internal;
+package com.liferay.portal.search.internal.osgi.util.tracker;
 
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.instance.lifecycle.PortalInstanceLifecycleListener;
